# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NetChat REST API for authentication, presence, users, and message history
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  online: boolean;
  /** @nullable */
  lastSeen: string | null;
}

export type MessageStatus = typeof MessageStatus[keyof typeof MessageStatus];


export const MessageStatus = {
  sent: 'sent',
  delivered: 'delivered',
  read: 'read',
} as const;

export interface Message {
  id: number;
  senderId: number;
  receiverId: number;
  message: string;
  timestamp: string;
  status: MessageStatus;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 24
     */
  username: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  displayName: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  /**
     * @minLength 3
     * @maxLength 24
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface AuthResponse {
  user: User;
}