# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NetChat REST API for authentication, presence, users, and message history
 * OpenAPI spec version: 0.1.0
 */
import type { MessageStatus } from './messageStatus';

export interface Message {
  id: number;
  senderId: number;
  receiverId: number;
  message: string;
  timestamp: Date;
  status: MessageStatus;
}