# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NetChat REST API for authentication, presence, users, and message history
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 24
     */
  username: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  displayName: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}