# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NetChat REST API for authentication, presence, users, and message history
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  displayName: string;
  online: boolean;
  /** @nullable */
  lastSeen: Date | null;
}